non-free/sgx-ddk-um: new aport (MR 1868)
From TI, they must not be redistributed in a modified form. They
are built against GLIBC 2.29

To be functional, they need to be paired with a special build
of Mesa.  There are subpackages for each different soc and only
one soc can be installed at a time.  The Exynos S5PV210 subpackage
uses the same SGX540-120 as the ti443x blobs, but needs a binary
edit to be loaded.  Due to the license restrictions, it must be
done in a post-install script.