main/postmarketos-ui-phosh: depend on xdg-desktop-portal-gtk (MR 1844)
Add the dependency to properly support file transfer between sandboxed
applications. Currently they can be used with flatpak (e.g. for quickly
testing out applications, not a recommended way to applications) and long
term we could use this in combination with apparmor for natively packaged
applications.

Notably the KDE version is already getting pulled in by when installing
postmarketos-ui-plasma-*.