device/pine64-pinephone: fix pogopin i2c bus (MR 1950)
The i2c bus for the pogo pins was locked up because the level shifter is
powered by the regulator used for the cameras. This enables that
specific regular for the gpios the i2c bus is on so the pull-ups work
when reading and writing from the pogo pins.

[ci:skip-build] Already built in CI