msm8974 devices: allow vram carveout (MR 2015)
This is needed with the 5.11 kernel. Also adjust the vram & cma sizes
for all msm8974 devices.

The CMA size is 256MB as set in the kernel config.

VRAM is default 16MB although it is set to 192MB with a kernel patch;
but set it manually for all devices to make it possible to remove this
patch in the future - until we have IOMMU working.