linux-purism-librem5: build led-pattern driver into kernel (MR 2043)
There's a race condition between when the pwmled device is initialized,
which kicks off a feedbackd udev rule that configures the led trigger to
'pattern', and when ledtrig_pattern actually loads. It seems to mostly
load after the rule runs, so the trigger is never set to 'pattern' on
boot.

This removes the race by building the ledtrig_pattern driver into the
kernel.

[ci:skip-build] Already build successfully in CI