community/asus-me176c: fix splash screen width and phosh scaling (MR 2047)
Actually the display in asus-me176c has a width of 800px, which means
that the splash screen isn't perfectly centered. Correct this in
deviceinfo to make it show up correctly.

Also, add configuration file for Phosh to reduce scaling a bit to 1.75.
It's kind of subjective but scaling = 2 definitely feels too large.

[ci:skip-build]: already built successfully in CI