device-oneplus-enchilada: add audio and fix osk-sdl (MR 2206)
Add the missing depends for our soc package alsa ucm configs. Also add
the missing depends to add the GPU firmware to the initramfs so that
osk-sdl works properly!

Also enable logs and point them to UART, this will clean up the splash
screen and let any developers with UART attached to debug the boot
process.