temp/upower: sync icon and warn for low battery level (MR 2255)
Commits cfb5d41b and ead17254 updated the default battery levels for
warnings and actions in the pinephone and the pinetab. Since then,
the chances of the phone dying for lack of battery power have completely
dissapeared. However, the battery icon in phosh was never updated to show
a "caution" battery level. The value was hardcoded and ignored the setting
in the config file. This commit fixes that issue