main/linux-postmarketos-exynos4: compress with xz (MR 2410)
gzip results in a boot.img that's too big (over 8 MB) for e.g.
samsung-i9300, so let's replace it with something that produces
smaller kernels. With this, the entire boot.img is 6,3 MB (down
from 8MiB, or 8.5458944 MB) which leaves plenty of space on the
8 MB partition and prevents the recovery from being overwritten
when this kernel is flashed.

Tested and works on samsung-i9300.

[ci:skip-build] already built successfully in CI