pine64-pinephone: fix pre-upgrade script (MR 2512)
Use -f, so rm
* does not complain if the file does not exist
* does not exit with 1 if the file does not exist

Remove the stdout/stderr redirect (not needed with -f).

Add exit 0 at the end of the file, even though it clearly should not
exit with anything other than 0 anymore, as it's common practice in
these pre-upgrade etc. scripts.

With the old script, the problem was that it would exit 1 as soon as the
files did not exist at the time the script runs. Exit 1 in this
pre-upgrade script causes apk to show an error, and it would not go away
with 'apk fix' since it ran into the same problem.

Fixes: 7c511d ("pine64-pinephone: remove old u-boot...")