samsung-gts210lte: new device (MR 2599)
Touchscreen and usb networking works.  Much of the hardware is the
same as on gts210velte (but that one has a qcom soc and not
exynos5433).  Exactly the same touchscreen calibrarion as for
gts210velte works for this device as well.

Also include firmware package with wifi files. Unlike on many other
Samsung devices wifi does not work out of the box after these
firmware files has been added.

[ci:skip-build] Already built on CI in MR