community/firmware-qcom-db410c: package all firmware, squash, new paths (MR 2619)
Unfortunately, at the moment MSM8916 devices always have to install
all of linux-firmware-qcom (~ 60 MiB) when they actually just use
~ 0.008 MiB of GPU firmware in there. Everything else is just big
firmware blobs for other SoCs that are never going to be useful.

It's much more efficient to package the necessary firmware separately
to save rootfs space. The DB410c linux-board-support-package contains
all the firmware from linux-firmware as well, so we can easily package
it as part of the firmware-qcom-db410c package.

This commit makes the following changes to firmware-qcom-db410c:
  1. Package venus() firmware that was previously installed from
     linux-firmware-qcom. GPU firmware is packaged in firmware-qcom-adreno.
  2. Use new firmware paths for DB410c that were applied upstream
  3. "Squash" the firmware blobs to a single file using pil-squasher