main/msm-firmware-loader: add dummy wcnss subpackage (MR 2619)
firmware-qcom-msm8916-wcnss is a virtual package used for MSM8916
devices without secure boot (i.e. device that do not require signed
firmware). For those it's best to install more recent firmware by
default rather than using the firmware on the firmware partition.

firmware-qcom-msm8916-wcnss is provided by firmware-qcom-db410c-wcnss
by default, however it's still useful to make it possible to switch to
the stock firmware on the firmware partition if needed. To allow that,
add a dummy msm-firmware-loader-wcnss package that provides the virtual
package using the msm-firmware-loader.

  # apk add msm-firmware-loader-wcnss
  (1/2) Purging firmware-qcom-db410c-wcnss (1034.2.1-r2)
  (2/2) Installing msm-firmware-loader-wcnss (1-r1)