fairphone-fp4: add firmware and sd card quirks (MR 2806)
With this you can get the modem and wifi running nearly out of the box.
As the kernel currently doesn't support UFS (internal storage) you need
to dump the following partitions using dd in TWRP or similar:

  fsc, fsg, modemst1, modemst2, tunning

Place those as files in the rootfs in /modem-partitions/ and add "-o
/modem-partitions" to the rmtfs arguments by modifying /etc/init.d/rmtfs
and changing command_args to "-P -r -o /modem-partitions".

[ci:skip-build] already built successfully in CI