device-purism-librem5: don't use legacy u-boot images for booting (MR 2842)
Using uInitrd is no longer necessary. I'm not sure *why*, but it just
works using the non-u-boot-image initramfs. The post-upgrade script was
changed to clean up the old images from /boot/u{Image,Initrd}, restoring
a lot of free space in that partition (~25MB)

This also sneaks in a small 1 character change to silence removing the
gpsd service from the default runlevel

[ci:skip-build]: already built successfully in CI