main/postmarketos-ui-xfce4: upgrade to 0.5.1 (MR 2935)
The main change is to move the postmarketOS-specific Xfce4 configs
from /etc/skel to /etc/xdg. This improves the config handling because
the location in /etc/xdg can be defined as the default fallback
configs.

Moving the gtk-3.0 folder to /etc/xdg did not work. Therefore this
one remains in /etc/skel.

Beyond this, there are some minor changes from the xfce4-phone
repository:
- Change icon path of whiskermenu button to new location
- Apply pmOS wallpaper to more monitors
- Several clean-ups in the config files