main/postmarketos-upgrade-kernel: fix custom boot partition being ignored (MR 3095)
This code wasn't actually working for me on samsung-m0, so I
investigated and came up with that this fixes it. I'm not sure if you
are supposed to be able to omit the deviceinfo_ prefix? Either way, it
doesn't seem to work, so let's fix it.

Fixes fad7c7ef8e4b47e054f6d571b451f310e12dbff4