temp/xkeyboard-config: upgrade to 2.35.1 (MR 3133)
Upgrade to latest version and add a backport of the PinePhone keyboard
patch by Undef, which is already merged to upstream master and therefore
will be in the next version.

Adjust the source URL to the same that is in the Alpine APKBUILD.

Keep the RX-51 (Nokia N900) patches for now, but it would be good if we
could upstream/drop them in order to use Alpine's package again in the
near future. Created pmaports issue 1521 for that.

Related: pmaports issue 1515
Co-Authored-By: Oliver Smith <ollieparanoid@postmarketos.org>