linux-postmarketos-exynos4: disable HDMI for Midas (MR 3395)
Disable HDMI as workaround, as it seems to be buggy when HDMI is
unplugged. It's apparently in DRM_MODE_CONNECTED state, but without size
and with no modes. According to Guido from Phosh this is a bug and
should be fixed in the kernel. By disabling this until it is fixed, we
get Phosh working again.

Related: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/828