cross/gcc-*: upgrade to 12.1.1_git20220630-r5
Removed isl as described in pmaports issue 2167. Verified locally that
gcc-aarch64 builds. Tried to test building packages with it, but wasn't
possible with the local build due to a bug, I'll create an issue and
link it in pmaports issue 2167. Pushing this now as it's a critical fix,
should work in theory as soon as the binary packages are built on bpo,
and right now it's broken anyway.