{main,cross}/gcc{4,6}*: remove !tracedeps option (MR 3498)
Don't build with !tracedeps, as otherwise the isl dependency is not
properly added to the built gccs. In practice this probably doesn't
matter since we only use these old gccs to build kernels in the native
chroots (not via crossdirect in foreign arch chroots), and we also
install a proper gcc in the native chroot that pulls in the isl
dependency. But fix it for correctness.

Related: pmaports issue 1732

[ci:skip-build] Already built successfuly on CI in MR