treewide: remove linux-headers replacements (MR 3597)
Remove the -dev subpackages of kernels that replace the linux-headers
package. As I understand, the only case where this would make sense is
if you wanted to build a kernel module against headers that need to be
newer than the headers in Alpine's linux-headers package (currently
5.19.5).

While this might have been historically relevant when wireguard wasn't
in the kernel and you might have needed newer linux-headers to build
the module, it seems irrelevant today. If somebody should need this in
the future, then bring it back properly / consider getting the
linux-headers package in Alpine upgraded. Right now the -dev packages
are broken in several ways:

* Use of unversioned provides= (pma#1766)
* Don't contain binaries needed to build binary packages (pma#462).
  This was worked around in 6aba5f ("Package kernel-scripts
  separately"), but later this didn't build anymore and we removed it.

So in short: broken and unused, drop them.