valve-jupiter: new device (MR 3877)
Note that only the "linux-edge" kernel is used, and it's likely that
this kernel is still missing some functionality. I'd like to avoid using
a downstream/forked kernel if at all possible...

I chose to use gummiboot for now, since it's really simple to configure.
This can be moved to a more "standard" EFI boot mechanism once pmOS
figures that out.

[ci:skip-build]: already built successfully in CI