main/postmarketos-bootsplash: kill pbsplash before display-manager (MR 3893)
If pbsplash is running when a display manager starts it often causes
issues as display managers want to run on a particular VT, but are
unable to switch the active VT since pbsplash has exclusive control of
the framebuffer.

Making DMs not start until after local would solve the issue but
it would also needlessly delay the boot process.

Until we have a better way to detect display managers starting and react
to it, this will at least get devices working again. Though due to how
openrc resolves service start order it may mean the splash is killed
earlier than necessary.