linux-google-sargo: remove (MR 4146)
It doesn't build currently (issue 2146) and also it's not used by any
device package currently (with other ports we support selecting between
mainline and downstream kernel in some cases, but not here).

If it's useful, the build can be fixed and it can be added back in a
future patch.