firmware-samsung-dream: add MFC firmware (MR 4188)
Otherwise there is a 3 min delay until we can ssh into the device.

Same issue as on samsung-crownlte, where rom4nik investigated and
found the solution [1].

Note that the display does not work for dreamlte and dream2lte at the
moment, the device gets stuck on the pmos boot splash screen. This
issue is tracked in
https://gitlab.com/postmarketOS/pmaports/-/issues/2102.

[1] https://gitlab.com/postmarketOS/pmaports/-/merge_requests/4176

[ci:skip-build] already built successfully in CI